Initial commit
Initial set up is an explorer page, which has been styled quite basically. The welcome screen also needs styling.

Still to add:
- Navigation buttons
- Welcome page stylesheet
- More areas adding

Next step should be to add another area next to Hyde Park, and allow travel between the two using the buttons.

Much later on I can think about adding  events in areas and adding a map in the footer with a 'you are here' feature on it.