Added notes and map overlay
I've added the notes and map overlay and placed them in the footer. This has replaced my original idea to have the map as an image at the bottom of the screen. I feel like this uses space better and looks nicer.

I've also added a sound to play when opening and closing the overlay.

The notes contains the case information first displayed on start up.