Added walking and first time message
I've updated the styling on the index. I've also added a delay when you click on a new html, and a walking sound effect plays.

Unfortunately, Chrome blocks favicons when the image is stored locally and blocks autoplaying of music, so I'm unable to have ambient sounds play when you enter a new area.

I've also added a message that displays the first time you enter an area.